asst 1/requirements.txt
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
jax==0.9.0.1
jax-cuda12-pjrt==0.9.0.1
jax-cuda12-plugin==0.9.0.1
jaxlib==0.9.0.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opt_einsum==3.4.0
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.0
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0+cu128
torchaudio==2.10.0+cu128
torchvision==0.25.0+cu128
triton==3.6.0
typing_extensions==4.15.0
